python-service/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pydantic==2.11.7
requests==2.32.4
python-dotenv==1.1.1